+++ packages/gatsby-cli/src/reporter/reporter.ts
@@ -7,6 +7,7 @@ import * as reporterActions from "./redux/actions"
 import { LogLevels, ActivityStatuses } from "./constants"
 import { getErrorFormatter } from "./errors"
 import constructError from "../structured-errors/construct-error"
+import { IErrorMapEntry, ErrorId } from "../structured-errors/error-map"
 import { prematureEnd } from "./catch-exit-signals"
 import { IStructuredError } from "../structured-errors/types"
 import { createTimerReporter, ITimerReporter } from "./reporter-timer"
@@ -35,6 +36,20 @@ class Reporter {
    */
   stripIndent = stripIndent
   format = chalk
+
+  errorMap: Record<ErrorId, IErrorMapEntry> = {}
+
+  /**
+   * Set a custom error map to the reporter. This allows
+   * the reporter to extend the internal error map
+   */
+
+  setErrorMap = (entry: Record<ErrorId, IErrorMapEntry>): void => {
+    this.errorMap = {
+      ...this.errorMap,
+      ...entry,
+    }
+  }
 
   /**
    * Toggle verbosity.
@@ -136,7 +151,8 @@ class Reporter {
       }
     }
 
-    const structuredError = constructError({ details })
+    const structuredError = constructError({ details }, this.errorMap)
+
     if (structuredError) {
       reporterActions.createLog(structuredError)
     }