+++ packages/gatsby-transformer-sqip/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.40](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@2.3.39...gatsby-transformer-sqip@2.3.40) (2020-10-08)
+
+**Note:** Version bump only for package gatsby-transformer-sqip
+
 ## [2.3.39](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@2.3.38...gatsby-transformer-sqip@2.3.39) (2020-10-06)
 
 **Note:** Version bump only for package gatsby-transformer-sqip